requirements.txt
protobuf
numpy
tensorflow
urllib3
tensorboard
requests
pyyaml
grpcio
